+++ CHANGES.md
@@ -37,6 +37,7 @@ This release exports all functions in `matplotlib.pyplot` to `latexplotlib`. Thi
 ## Version 0.x.x
 
 - update `lpl.subplots` docstring
+- typing: allow both floats and ints as widths and height
 
 ### Development
 - remove unused packages from environment.yml